Expr

Description

Handles the parsing of user input and finds the values of expressions given in this input

Documentation

type Name = String

Type Name is just an alias for String

type File = Name

Type File is just an alias for Name

data Expr

Data type Expr which holds the possible signatures for each of the calculator operations

Constructors

Add Expr Expr
Addition of two Integers/Variables

Val Int
Retrieving value of an entry (e.g. input is just "4")

Name [Char]
Retrieving value of variable

Minus Expr Expr
Subtraction of two Integers/Variables

Multiply Expr Expr
Multiplication of two Integers/Variables

Division Expr Expr
Division of two Integers/Variables

Abs Expr
Find the Absolute value of an expression

Mod Expr Expr
Find the Modulus of two Integers/Variables

Power Expr Expr
Find the Power of one IntegerVariable to another IntegerVariable

data Command

Data type for the possible REPL commands and the signatures it can have

Constructors

Set Name Expr
Variable Assignment

Eval Expr
Evaluation of an expression

Quit
Quit Command

Read File
Read file Command

retrieveVar :: [(Name, Int)] -> Name -> Maybe Int

Function that recursively goes through list of variables in order to find value of a variable, or returns nothing if that variable doesn't exist

eval :: [(Name, Int)] -> Expr -> Maybe Int

eval function takes in the current list of variables, the expression to be evaluated and returns a Maybe Int (in case the expression fails). This is the framework for all the caclulations the calculator can make.

digitToInt :: Char -> Int

Turns a Char into an Int assuming the Char is a valid Int

pCommand :: Parser Command

Function that finds what type of command the input given by the user is (Evaluation, Variable Assignment, Quit or File Read )

pExpr :: Parser Expr

Function to handle simple expressions such as Add and Minus that have the lowest precedence

pFactor :: Parser Expr

Function to handle some of the complex expressions that have higher precedence such as those containing brackets or Abs, or those that have to be checked first such as getting the value of variables

pTerm :: Parser Expr

Function for the normal expressions that have a medium level of precedence, including the Multiply, Division, Power and Mod expressions