REPL

Description

Handles overarching repl loop of the program and State operations such as updating History or removing variables.

Documentation

data State

State is a data type that stores the history of commands in the program and the current value of any variables

Constructors

State
Generic State Object

Fields

vars :: [(Name, Int)]

Variable History

history :: [Command]

History Tracking

initState :: State

Creation of empty State object

updateVars :: Name -> Int -> [(Name, Int)] -> [(Name, Int)]

Takes in the name of a variable and the new value of the variable, plus a list of the current variables. Will remove any variables that currently exist within the list via dropVar and then return the new list of variables

dropVar :: Name -> [(Name, Int)] -> [(Name, Int)]

Return a new set of variables with the given name removed using the current list passed in by updateVars

addHistory :: State -> Command -> State

Add a command to the command history in the State

checkLength :: Int -> [Command] -> Bool

Checks the length of a list (-1) is more than the index trying to be accessed and returns a Bool result

historyCheck :: State -> [Char] -> IO ()

Checks that the point being accessed in the state's history is valid and if so passes the command from that point into the process function

checkDigits :: [Char] -> Bool

Recursively loops through the list of Chars to check if they are all digits or not, and then returns a Bool value

processLine :: State -> Command -> [String] -> IO ()

Another version of process that is used for file reading as this one loops back into the processFile method rather than going on to the repl loop, allowing the program to read files with multiple lines

parseOpr :: State -> String -> [String] -> IO ()

Special version of repl that is used to parse the operations of a read in file that will not loop back on itself if there is an error

processFile :: State -> [String] -> IO ()

Function that is used to process the list of commands read in from a file

process :: State -> Command -> IO ()

Process function that has multiple different results depending on the type of command that has been entered that checks if the command is valid or not according to the eval function and then performs the correct operation depending on the results of eval

repl :: State -> IO ()

Read, Eval, Print Loop This reads and parses the input using the pCommand parser, and calls process to process the command. process will call repl when done, so the system loops.